virtualDocument class
